src/requirements.txt
functions-framework==3.5.0
stripe==7.14.0
google-cloud-secret-manager==2.17.0
google-cloud-bigquery==3.12.0
